+++ Web/Program.cs
@@ -44,7 +44,9 @@ builder.Services
         s.Title = "FastEndpoints Sandbox";
         s.Version = "v2.0";
     })
-    .AddSwaggerDoc(minEndpointVersion: 3, maxEndpointVersion: 3, settings: s =>
+
+    //only ver3 & only FastEndpoints
+    .AddSwaggerDoc(minEndpointVersion: 3, maxEndpointVersion: 3, excludeNonFastEndpoints: true, settings: s =>
     {
         s.DocumentName = "Release 3.0";
         s.Title = "FastEndpoints Sandbox ver3 only";